backend/requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-noise-cancellation
livekit-plugins-google
requests
python-dotenv
pydantic-ai-slim[openai,mcp]
livekit-agents[tavus]~=1.0
tzdata
flask[async]
flask
flask-cors